CAMEL-17894: move some of the JavaDoc code to their own utility class
